requirements.txt
Django==5.0
django-crispy-forms==2.1
Pillow==10.1.0
python-dotenv==1.0.0
django-bootstrap5==23.3
django-allauth==0.59.0
python-telegram-bot==13.7
celery==5.3.0
redis==4.5.0
django-celery-beat==2.5.0